@@ -1,6 +1,7 @@
 {
   "name": "http-proxy-middleware",
-  "version": "3.0.1-beta.1",
+  "type": "commonjs",
+  "version": "3.0.1",
   "description": "The one-liner node.js proxy middleware for connect, express, next.js and more",
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
@@ -12,7 +13,7 @@
     "install:all": "yarn && (cd examples && yarn)",
     "lint": "yarn prettier && yarn eslint",
     "lint:fix": "yarn prettier:fix && yarn eslint:fix",
-    "eslint": "eslint '{src,test}/**/*.ts' --cache",
+    "eslint": "eslint '{src,test,examples}/**/*.{js,ts}' --cache",
     "eslint:fix": "yarn eslint --fix",
     "prettier": "prettier --list-different \"**/*.{js,ts,md,yml,json,html}\"",
     "prettier:fix": "prettier --write \"**/*.{js,ts,md,yml,json,html}\"",
@@ -55,42 +56,45 @@
   },
   "homepage": "https://github.com/chimurai/http-proxy-middleware#readme",
   "devDependencies": {
-    "@commitlint/cli": "19.3.0",
-    "@commitlint/config-conventional": "19.2.2",
+    "@commitlint/cli": "19.4.1",
+    "@commitlint/config-conventional": "19.4.1",
+    "@eslint/js": "9.9.1",
     "@types/debug": "4.1.12",
+    "@types/eslint": "9.6.1",
+    "@types/eslint__js": "8.42.3",
     "@types/express": "4.17.21",
     "@types/is-glob": "4.0.4",
     "@types/jest": "29.5.12",
     "@types/micromatch": "4.0.9",
-    "@types/node": "20.14.10",
+    "@types/node": "22.5.1",
     "@types/supertest": "6.0.2",
-    "@types/ws": "8.5.10",
-    "@typescript-eslint/eslint-plugin": "7.16.0",
-    "@typescript-eslint/parser": "7.16.0",
+    "@types/ws": "8.5.12",
     "body-parser": "1.20.2",
-    "eslint": "8.57.0",
+    "eslint": "9.9.1",
     "eslint-config-prettier": "9.1.0",
-    "eslint-plugin-prettier": "5.1.3",
+    "eslint-plugin-prettier": "5.2.1",
     "express": "4.19.2",
     "get-port": "5.1.1",
-    "husky": "9.0.11",
+    "globals": "15.9.0",
+    "husky": "9.1.5",
     "jest": "29.7.0",
-    "lint-staged": "15.2.7",
-    "mockttp": "3.14.0",
+    "lint-staged": "15.2.9",
+    "mockttp": "3.15.2",
     "open": "8.4.2",
-    "prettier": "3.3.2",
+    "prettier": "3.3.3",
     "supertest": "7.0.0",
-    "ts-jest": "29.2.2",
-    "typescript": "5.5.3",
+    "ts-jest": "29.2.5",
+    "typescript": "5.5.4",
+    "typescript-eslint": "8.3.0",
     "ws": "8.18.0"
   },
   "dependencies": {
-    "@types/http-proxy": "^1.17.14",
-    "debug": "^4.3.5",
+    "@types/http-proxy": "^1.17.15",
+    "debug": "^4.3.6",
     "http-proxy": "^1.18.1",
     "is-glob": "^4.0.3",
-    "is-plain-obj": "^3.0.0",
-    "micromatch": "^4.0.7"
+    "is-plain-object": "^5.0.0",
+    "micromatch": "^4.0.8"
   },
   "engines": {
     "node": "^14.15.0 || ^16.10.0 || >=18.0.0"
