@@ -1,4 +1,4 @@
-/* eslint-disable @typescript-eslint/ban-types */
+/* eslint-disable @typescript-eslint/no-unsafe-function-type */
 
 export function getFunctionName(fn: Function): string {
   return fn.name || '[anonymous Function]';
